[Refactor] #284 Settings화면 유닛 테스트 수정
