+++ SanTa/SettingsSceneTests/SettingsSceneTests.swift
@@ -1,5 +1,5 @@
 //
-//  SettingsUseCaseTests.swift
+//  SettingsViewModelTests.swift
 //  SettingsTests
 //
 //  Created by CHANGMIN OH on 2021/11/08.
@@ -7,7 +7,7 @@
 
 import XCTest
 
-class SettingsUsecaseTests: XCTestCase {
+class SettingsViewModelTests: XCTestCase {
     
     class MockRepository: SettingsRepository {
         func save<T>(value: T, key: Settings) where T : Decodable, T : Encodable {
@@ -23,30 +23,38 @@ class SettingsUsecaseTests: XCTestCase {
         }
     }
     
+    private var viewModel: SettingsViewModel!
     private var useCase: SettingsUsecase!
-    private var repository: MockRepository!
     
     override func setUpWithError() throws {
-        repository = MockRepository()
-        useCase = DefaultSettingsUsecase(settingsRepository: repository)
+        useCase = SettingsUsecase(settingsRepository: MockRepository())
+        viewModel = SettingsViewModel(settingsUseCase: useCase)
     }
     
-    func test_repository반환_값에_따른_배열생성() {
-        let options = useCase.makeSettings()
+    func test_ViewModel은_viewDidLoad시_UseCase의_반환값_settings프로퍼티에_세팅() throws {
+        viewModel.viewDidLoad()
+        XCTAssertEqual(viewModel.sectionCount, 3)
+    }
+    
+    func test_ViewModel은_change호출시_문자열이면_settings프로퍼티_업데이트() throws {
+        viewModel.change(value: 1, key: .mapFormat)
+        XCTAssertEqual(viewModel.sectionCount, 0)
         
+        viewModel.change(value: "string", key: .mapFormat)
+        XCTAssertEqual(viewModel.sectionCount, 3)
+    }
+    
+    func test_UseCase는_repository반환_값에_따른_배열생성() {
+        let options = useCase.makeSettings()
+
         let option1 = options[0][0] as? ToggleOption
         let option2 = options[0][1] as? ToggleOption
         let option3 = options[1][0] as? ToggleOption
-        let option4 = options[1][1] as? ToggleOption
-        let option5 = options[2][0] as? ToggleOption
-        let option6 = options[3][0] as? MapOption
-        
+        let option4 = options[2][0] as? MapOption
+
         XCTAssertNotNil(option1)
         XCTAssertNotNil(option2)
         XCTAssertNotNil(option3)
         XCTAssertNotNil(option4)
-        XCTAssertNotNil(option5)
-        XCTAssertNotNil(option6)
     }
 }
-
